@@ -1,10 +1,15 @@
-[tool.poetry]
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
 name = "dbt-osmosis"
-version = "0.14.0"
-description = "A dbt server and suite of optional developer tools to make developing with dbt delightful."
-authors = ["z3z1ma <butler.alex2010@gmail.com>"]
-license = "Apache-2.0"
+version = "1.0.0"
+description = "A dbt utility for managing YAML to make developing with dbt more delightful."
 readme = "README.md"
+license = { text = "Apache-2.0" }
+authors = [{ name = "z3z1ma", email = "butler.alex2010@gmail.com" }]
+keywords = ["dbt", "yaml", "osmosis"]
 classifiers = [
   "Development Status :: 4 - Beta",
   "Intended Audience :: Developers",
@@ -16,66 +21,49 @@ classifiers = [
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
 ]
-keywords = ["dbt", "server", "streamlit", "git", "diff"]
-documentation = "https://github.com/z3z1ma/dbt-osmosis"
-repository = "https://github.com/z3z1ma/dbt-osmosis"
+requires-python = ">=3.9,<3.13"
 
-[tool.poetry.dependencies]
-python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
-click = ">7"
-dbt-core = ">=1.8,<1.10"
-"ruamel.yaml" = ">=0.17"
-rich = ">=10"
-GitPython = ">3,<4"
-# Streamlit Workbench Dependencies
-streamlit = { version = ">=1.20.0", optional = true }
-streamlit-ace = { version = ">=0.1.0", optional = true }
-ydata-profiling = { version = ">=3.6.0", optional = true }
-feedparser = { version = "^6.0.10", optional = true }
-streamlit-elements-fluence = { version = ">=0.1.4", optional = true }
-# Testing duckdb + sqlite
-dbt-duckdb = { version = ">=1.0.0", optional = true }
-dbt-sqlite = { version = ">=1.0.0", optional = true }
-dbt-postgres = { version = ">=1.0.0", optional = true }
-openai = { version = ">0.28.0", optional = true }
-
-[tool.poetry.extras]
-duckdb = ["dbt-duckdb"]
-sqlite = ["dbt-sqlite"]
-postgres = ["dbt-postgres"]
-workbench = [
-  "streamlit",
-  "streamlit-ace",
-  "ydata-profiling",
-  "feedparser",
-  "streamlit-elements-fluence",
+dependencies = [
+  "click>7,<9",
+  "dbt-core>=1.8,<=1.10",
+  "ruamel.yaml~=0.18.7",
+  "rich>=10",
+  "pluggy>=1.5.0,<2",
+  "typing-extensions~=4.12.2 ; python_version < '3.10'",
 ]
 
-[tool.poetry.group.dev.dependencies]
-black = ">=21.9b0"
-mypy = ">=0.910"
-pylint = ">=2.11.1"
-pytest = "^7.4.2"
-pre-commit = ">3.0.0"
+[project.optional-dependencies]
+duckdb = ["dbt-duckdb>=1.0.0,<2"]
+sqlite = ["dbt-sqlite>=1.0.0,<2"]
+postgres = ["dbt-postgres>=1.0.0,<2"]
 
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+workbench = [
+  "streamlit>=1.20.0,<1.36.0",
+  "streamlit-ace~=0.1.1",
+  "ydata-profiling~=4.12.1",
+  "feedparser~=6.0.11",
+  "streamlit-elements-fluence>=0.1.4",
+  "setuptools>=70",
+]
 
-[tool.black] # https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html#configuration-via-a-file
+openai = ["openai>0.28.0"]
+
+dev = ["ruff~=0.8.4", "pytest~=8.3.4", "pre-commit>3.0.0"]
+
+[project.scripts]
+"dbt-osmosis" = "dbt_osmosis.cli.main:cli"
+
+[tool.black]
 line-length = 100
-target-version = ["py38", "py39", "py310", "py311"]
+target-version = ["py39", "py310", "py311", "py312"]
 preview = true
 
-[tool.isort] # https://pycqa.github.io/isort/docs/configuration/options.html
+[tool.isort]
 color_output = true
 line_length = 100
 profile = "black"
 src_paths = ["src"]
 include_trailing_comma = true
 
-[tool.ruff] # https://beta.ruff.rs/docs/configuration/#using-pyprojecttoml
+[tool.ruff]
 line-length = 100
-
-[tool.poetry.scripts]
-dbt-osmosis = 'dbt_osmosis.main:cli'
